Zot is now the final boss.
The final boss does not need to be the hardest boss in the game,
because some people will be playing a 3-rune game and some will
be playing a 15-rune game.
It would be great if the final boss was more interesting than
most.
To that end, Zot:5 now has a thrown room. You have to kill Zot
before you can take her heart. Also, we have a mock thrown room
vault that can pop up elsewhere in Zot.